Fixed bugs in plotting and feature selection cells
The first was where I had thought I was changing axes labels but was actually
changing variables names. This is now reverted.

The second is where the code would try to examine different numbers of features
which could exceed the number of features in the dataset. I added a filter to
fix this.